+++ src/app/(default)/layout.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from 'react'
-import { Header } from '../_components/header'
-import { Footer } from '../_components/footer'
+import { Header } from '~/components/header'
+import { Footer } from '~/components/footer'
 
 export default async function DefaultLayout({
   children,
