+++ src/components/post-summary.tsx
@@ -1,11 +1,11 @@
 import Link from 'next/link'
 import { summarize } from '~/server/summary'
 import { type RouterOutputs } from '~/trpc/shared'
-import { Banner } from '~/app/_components/banner'
+import { Banner } from '~/components/banner'
 import { classNames } from '~/utils/core'
 import { formatDistanceToNow } from 'date-fns/formatDistanceToNow'
-import { AuthorWithDate } from '~/app/_components/author-with-date'
-import { HtmlView } from '~/app/_components/html-view'
+import { AuthorWithDate } from '~/components/author-with-date'
+import { HtmlView } from '~/components/html-view'
 import ChevronRightIcon from '~/components/svg/chevron-right-icon'
 import MessageIcon from '~/components/svg/message-icon'
