+++ src/components/post-view.tsx
@@ -4,7 +4,7 @@ import { api } from '~/trpc/react'
 import { type RouterOutputs } from '~/trpc/shared'
 import { Banner } from './banner'
 import { useSession } from 'next-auth/react'
-import { PostAction } from '../(default)/post/[id]/_components/post-action'
+import { PostAction } from '~/components/post-action'
 import { AuthorWithDate } from './author-with-date'
 import { HtmlView } from './html-view'
 import { ReactionButton } from './reaction-button'
